download account metadata and small vochain refactor
The offchain data downloader is now subscribed to setAccount events in order
to download the metadata (if exist) of an account.

Small refactor on vochian package.
- Prefix with "state_" the files that contain state related methods.
- Split in more files
- Move some functions to transactions.go (those that perform tx checks)
