+++ vochain/state.go
@@ -2,27 +2,23 @@ package vochain
 
 import (
 	"bytes"
-	"crypto/sha256"
 	"encoding/binary"
 	"errors"
 	"fmt"
 	"os"
 	"path"
 	"path/filepath"
-	"sync"
 	"sync/atomic"
 	"time"
 
 	"github.com/ethereum/go-ethereum/common"
 	lru "github.com/hashicorp/golang-lru"
 	"github.com/vocdoni/arbo"
-	"go.vocdoni.io/dvote/crypto/ethereum"
 	"go.vocdoni.io/dvote/db"
 	"go.vocdoni.io/dvote/db/metadb"
 	"go.vocdoni.io/dvote/log"
 	"go.vocdoni.io/dvote/statedb"
 
-	"go.vocdoni.io/dvote/types"
 	"go.vocdoni.io/proto/build/go/models"
 	"google.golang.org/protobuf/proto"
 )
@@ -36,171 +32,7 @@ const (
 	storageDirectory = "vcstate"
 	// snapshotsDirectory is the final directory name where the snapshots are stored.
 	snapshotsDirectory = "snapshots"
-	// treasurerKey is the key representing the Treasurer entry on the Extra subtree
-	treasurerKey = "treasurer"
 )
-
-// TxTypeCostToStateKey translates models.TxType to a string which the State uses
-// as a key internally under the Extra tree
-var (
-	TxTypeCostToStateKey = map[models.TxType]string{
-		models.TxType_SET_PROCESS_STATUS:         "c_setProcessStatus",
-		models.TxType_SET_PROCESS_CENSUS:         "c_setProcessCensus",
-		models.TxType_SET_PROCESS_QUESTION_INDEX: "c_setProcessResults",
-		models.TxType_SET_PROCESS_RESULTS:        "c_setProcessQuestionIndex",
-		models.TxType_REGISTER_VOTER_KEY:         "c_registerKey",
-		models.TxType_NEW_PROCESS:                "c_newProcess",
-		models.TxType_SEND_TOKENS:                "c_sendTokens",
-		models.TxType_SET_ACCOUNT_INFO_URI:       "c_setAccountInfoURI",
-		models.TxType_CREATE_ACCOUNT:             "c_createAccount",
-		models.TxType_ADD_DELEGATE_FOR_ACCOUNT:   "c_addDelegateForAccount",
-		models.TxType_DEL_DELEGATE_FOR_ACCOUNT:   "c_delDelegateForAccount",
-		models.TxType_COLLECT_FAUCET:             "c_collectFaucet",
-	}
-	ErrProcessChildLeafRootUnknown = fmt.Errorf("process child leaf root is unkown")
-	ErrTxCostNotFound              = fmt.Errorf("transaction cost is not set")
-)
-
-// rootLeafGetRoot is the GetRootFn function for a leaf that is the root
-// itself.
-func rootLeafGetRoot(value []byte) ([]byte, error) {
-	if len(value) != defaultHashLen {
-		return nil, fmt.Errorf("len(value) = %v != %v: %w",
-			len(value),
-			defaultHashLen,
-			ErrProcessChildLeafRootUnknown)
-	}
-	return value, nil
-}
-
-// rootLeafSetRoot is the SetRootFn function for a leaf that is the root
-// itself.
-func rootLeafSetRoot(value []byte, root []byte) ([]byte, error) {
-	if len(value) != defaultHashLen {
-		return nil, fmt.Errorf("len(value) = %v != %v", len(value), defaultHashLen)
-	}
-	return root, nil
-}
-
-// processGetCensusRoot is the GetRootFn function to get the rolling census
-// root of a process leaf.
-func processGetCensusRoot(value []byte) ([]byte, error) {
-	var sdbProc models.StateDBProcess
-	if err := proto.Unmarshal(value, &sdbProc); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal StateDBProcess: %w", err)
-	}
-	if len(sdbProc.Process.RollingCensusRoot) != defaultHashLen {
-		return nil, fmt.Errorf("len(sdbProc.Process.RollingCensusRoot) != %v: %w",
-			defaultHashLen,
-			ErrProcessChildLeafRootUnknown)
-	}
-	return sdbProc.Process.RollingCensusRoot, nil
-}
-
-// processSetCensusRoot is the SetRootFn function to set the rolling census
-// root of a process leaf.
-func processSetCensusRoot(value []byte, root []byte) ([]byte, error) {
-	var sdbProc models.StateDBProcess
-	if err := proto.Unmarshal(value, &sdbProc); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal StateDBProcess: %w", err)
-	}
-	sdbProc.Process.RollingCensusRoot = root
-	newValue, err := proto.Marshal(&sdbProc)
-	if err != nil {
-		return nil, fmt.Errorf("cannot marshal StateDBProcess: %w", err)
-	}
-	return newValue, nil
-}
-
-// processGetPreRegisterNullifiersRoot is the GetRootFn function to get the nullifiers
-// root of a process leaf.
-func processGetPreRegisterNullifiersRoot(value []byte) ([]byte, error) {
-	var sdbProc models.StateDBProcess
-	if err := proto.Unmarshal(value, &sdbProc); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal StateDBProcess: %w", err)
-	}
-	if len(sdbProc.Process.NullifiersRoot) != defaultHashLen {
-		return nil, fmt.Errorf("len(sdbProc.Process.NullifiersRoot) != %v: %w",
-			defaultHashLen,
-			ErrProcessChildLeafRootUnknown,
-		)
-	}
-	return sdbProc.Process.NullifiersRoot, nil
-}
-
-// processSetPreRegisterNullifiersRoot is the SetRootFn function to set the nullifiers
-// root of a process leaf.
-func processSetPreRegisterNullifiersRoot(value []byte, root []byte) ([]byte, error) {
-	var sdbProc models.StateDBProcess
-	if err := proto.Unmarshal(value, &sdbProc); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal StateDBProcess: %w", err)
-	}
-	sdbProc.Process.NullifiersRoot = root
-	newValue, err := proto.Marshal(&sdbProc)
-	if err != nil {
-		return nil, fmt.Errorf("cannot marshal StateDBProcess: %w", err)
-	}
-	return newValue, nil
-}
-
-// processGetVotesRoot is the GetRootFn function to get the votes root of a
-// process leaf.
-func processGetVotesRoot(value []byte) ([]byte, error) {
-	var sdbProc models.StateDBProcess
-	if err := proto.Unmarshal(value, &sdbProc); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal StateDBProcess: %w", err)
-	}
-	if len(sdbProc.VotesRoot) != defaultHashLen {
-		return nil, fmt.Errorf(
-			"len(sdbProc.VotesRoot) != %v: %w",
-			defaultHashLen,
-			ErrProcessChildLeafRootUnknown)
-	}
-	return sdbProc.VotesRoot, nil
-}
-
-// processSetVotesRoot is the SetRootFn function to set the votes root of a
-// process leaf.
-func processSetVotesRoot(value []byte, root []byte) ([]byte, error) {
-	var sdbProc models.StateDBProcess
-	if err := proto.Unmarshal(value, &sdbProc); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal StateDBProcess: %w", err)
-	}
-	sdbProc.VotesRoot = root
-	newValue, err := proto.Marshal(&sdbProc)
-	if err != nil {
-		return nil, fmt.Errorf("cannot marshal StateDBProcess: %w", err)
-	}
-	return newValue, nil
-}
-
-// EventListener is an interface used for executing custom functions during the
-// events of the block creation process.
-// The order in which events are executed is: Rollback, OnVote, Onprocess, On..., Commit.
-// The process is concurrency safe, meaning that there cannot be two sequences
-// happening in parallel.
-//
-// If Commit() returns ErrHaltVochain, the error is considered a consensus
-// failure and the blockchain will halt.
-//
-// If OncProcessResults() returns an error, the results transaction won't be included
-// in the blockchain. This event relays on the event handlers to decide if results are
-// valid or not since the Vochain State do not validate results.
-type EventListener interface {
-	OnVote(vote *models.Vote, voterID VoterID, txIndex int32)
-	OnNewTx(hash []byte, blockHeight uint32, txIndex int32)
-	OnProcess(pid, eid []byte, censusRoot, censusURI string, txIndex int32)
-	OnProcessStatusChange(pid []byte, status models.ProcessStatus, txIndex int32)
-	OnCancel(pid []byte, txIndex int32)
-	OnProcessKeys(pid []byte, encryptionPub string, txIndex int32)
-	OnRevealKeys(pid []byte, encryptionPriv string, txIndex int32)
-	OnProcessResults(pid []byte, results *models.ProcessResult, txIndex int32) error
-	OnProcessesStart(pids [][]byte)
-	OnSetAccount(addr []byte, account *Account) error
-	OnTransferTokens(from, to []byte, amount uint64) error
-	Commit(height uint32) (err error)
-	Rollback()
-}
 
 type ErrHaltVochain struct {
 	reason error
@@ -208,13 +40,6 @@ type ErrHaltVochain struct {
 
 func (e ErrHaltVochain) Error() string { return fmt.Sprintf("halting vochain: %v", e.reason) }
 func (e ErrHaltVochain) Unwrap() error { return e.reason }
-
-// treeTxWithMutex is a wrapper over TreeTx with a mutex for convenient
-// RWLocking.
-type treeTxWithMutex struct {
-	*statedb.TreeTx
-	sync.RWMutex
-}
 
 // State represents the state of the vochain application
 type State struct {
@@ -367,35 +192,6 @@ func (v *State) SetChainID(chID string) {
 	v.chainID = chID
 }
 
-// MainTreeView is a thread-safe function to obtain a pointer to the last
-// opened mainTree as a TreeView.
-func (v *State) MainTreeView() *statedb.TreeView {
-	return v.mainTreeViewValue.Load().(*statedb.TreeView)
-}
-
-// setMainTreeView is a thread-safe function to store a pointer to the last
-// opened mainTree as TreeView.
-func (v *State) setMainTreeView(treeView *statedb.TreeView) {
-	v.mainTreeViewValue.Store(treeView)
-}
-
-// mainTreeViewer returns the mainTree as a treeViewer.
-// When committed is false, the mainTree returned is the not yet commited one
-// from the currently open StateDB transaction.
-// When committed is true, the mainTree returned is the last commited version.
-func (v *State) mainTreeViewer(committed bool) statedb.TreeViewer {
-	if committed {
-		return v.MainTreeView()
-	}
-	return v.Tx.AsTreeView()
-}
-
-// AddEventListener adds a new event listener, to receive method calls on block
-// events as documented in EventListener.
-func (v *State) AddEventListener(l EventListener) {
-	v.eventListeners = append(v.eventListeners, l)
-}
-
 var exist = []byte{1}
 
 // AddOracle adds a trusted oracle given its address if not exists
@@ -464,172 +260,6 @@ func (v *State) IsOracle(addr common.Address) (bool, error) {
 		}
 		return false
 	}(), nil
-}
-
-// SetTreasurer saves the Treasurer address to the state
-func (v *State) SetTreasurer(address common.Address, nonce uint32) error {
-	tBytes, err := proto.Marshal(
-		&models.Treasurer{
-			Address: address.Bytes(),
-			Nonce:   nonce,
-		},
-	)
-	if err != nil {
-		return err
-	}
-	v.Tx.Lock()
-	defer v.Tx.Unlock()
-	return v.Tx.DeepSet([]byte(treasurerKey), tBytes, StateTreeCfg(TreeExtra))
-}
-
-// Treasurer returns the address and the Treasurer nonce
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) Treasurer(committed bool) (*models.Treasurer, error) {
-	if !committed {
-		v.Tx.RLock()
-		defer v.Tx.RUnlock()
-	}
-	extraTree, err := v.mainTreeViewer(committed).SubTree(StateTreeCfg(TreeExtra))
-	if err != nil {
-		return nil, err
-	}
-	var rawTreasurer []byte
-	if rawTreasurer, err = extraTree.Get([]byte(treasurerKey)); err != nil {
-		return nil, err
-	}
-	var t models.Treasurer
-	if err := proto.Unmarshal(rawTreasurer, &t); err != nil {
-		return nil, err
-	}
-	return &t, nil
-}
-
-// IsTreasurer returns true if the given address matches the Treasurer address
-func (v *State) IsTreasurer(addr common.Address) (bool, error) {
-	t, err := v.Treasurer(false)
-	if err != nil {
-		return false, err
-	}
-	return addr == common.BytesToAddress(t.Address), nil
-}
-
-// IncrementTreasurerNonce increments the treasurer nonce
-func (v *State) IncrementTreasurerNonce() error {
-	t, err := v.Treasurer(false)
-	if err != nil {
-		return fmt.Errorf("incrementTreasurerNonce(): %w", err)
-	}
-	v.Tx.Lock()
-	defer v.Tx.Unlock()
-	t.Nonce++
-	tBytes, err := proto.Marshal(t)
-	if err != nil {
-		return fmt.Errorf("incrementTreasurerNonce(): %w", err)
-	}
-	log.Debugf("incrementing treasurer nonce, new nonce is %d", t.Nonce)
-	return v.Tx.DeepSet([]byte(treasurerKey), tBytes, StateTreeCfg(TreeExtra))
-}
-
-// SetTxCost sets the given transaction cost
-func (v *State) SetTxCost(txType models.TxType, cost uint64) error {
-	key, ok := TxTypeCostToStateKey[txType]
-	if !ok {
-		return fmt.Errorf("txType %v shouldn't cost anything", txType)
-	}
-	v.Tx.Lock()
-	defer v.Tx.Unlock()
-	log.Debugf("setting tx cost %d for tx %s", cost, txType.String())
-	costBytes := make([]byte, 8)
-	binary.LittleEndian.PutUint64(costBytes[:], cost)
-	return v.Tx.DeepSet([]byte(key), costBytes[:], StateTreeCfg(TreeExtra))
-}
-
-// TxCost returns the cost of a given transaction
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) TxCost(txType models.TxType, committed bool) (uint64, error) {
-	key, ok := TxTypeCostToStateKey[txType]
-	if !ok {
-		return 0, fmt.Errorf("txType %v shouldn't cost anything", txType)
-	}
-	if !committed {
-		v.Tx.RLock()
-		defer v.Tx.RUnlock()
-	}
-	extraTree, err := v.mainTreeViewer(committed).SubTree(StateTreeCfg(TreeExtra))
-	if err != nil {
-		return 0, err
-	}
-	var cost []byte
-	if cost, err = extraTree.Get([]byte(key)); err != nil {
-		return 0, ErrTxCostNotFound
-	}
-	return binary.LittleEndian.Uint64(cost), nil
-}
-
-// FaucetNonce returns true if the key is found in the subtree
-// key == hash(address, nonce)
-// committed is relative to the state on which the function is executed
-func (v *State) FaucetNonce(key []byte, committed bool) (bool, error) {
-	if !committed {
-		v.Tx.RLock()
-		defer v.Tx.RUnlock()
-	}
-	faucetNonceTree, err := v.mainTreeViewer(committed).SubTree(StateTreeCfg(TreeFaucet))
-	if err != nil {
-		return false, err
-	}
-	var found bool
-	if err := faucetNonceTree.Iterate(
-		func(k, _ []byte) bool {
-			if bytes.Equal(key, k) {
-				found = true
-				return true
-			}
-			return true
-		},
-	); err != nil {
-		return false, err
-	}
-	return found, nil
-}
-
-// SetFaucetNonce stores an already used faucet nonce in the
-// FaucetNonce subtree
-func (v *State) SetFaucetNonce(key []byte) error {
-	v.Tx.Lock()
-	defer v.Tx.Unlock()
-	return v.Tx.DeepSet(key, nil, StateTreeCfg(TreeFaucet))
-}
-
-// hexPubKeyToTendermintEd25519 decodes a pubKey string to a ed25519 pubKey
-/*
-func hexPubKeyToTendermintEd25519(pubKey string) (tmcrypto.PubKey, error) {
-	var tmkey ed25519.PubKey
-	pubKeyBytes, err := hex.DecodeString(pubKey)
-	if err != nil {
-		return nil, err
-	}
-	if len(pubKeyBytes) != 32 {
-		return nil, fmt.Errorf("pubKey length is invalid")
-	}
-	copy(tmkey[:], pubKeyBytes[:])
-	return tmkey, nil
-}
-*/
-
-// AddValidator adds a tendemint validator if it is not already added
-func (v *State) AddValidator(validator *models.Validator) error {
-	v.Tx.Lock()
-	defer v.Tx.Unlock()
-	validatorBytes, err := proto.Marshal(validator)
-	if err != nil {
-		return err
-	}
-	return v.Tx.DeepSet(validator.Address, validatorBytes, StateTreeCfg(TreeValidators))
 }
 
 // RemoveValidator removes a tendermint validator identified by its address
@@ -741,218 +371,6 @@ func (v *State) RevealProcessKeys(tx *models.AdminTx) error {
 		l.OnRevealKeys(tx.ProcessId, ekey, v.TxCounter())
 	}
 	return nil
-}
-
-// VoteCount return the global vote count.
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) VoteCount(committed bool) (uint64, error) {
-	if !committed {
-		v.Tx.RLock()
-		defer v.Tx.RUnlock()
-	}
-	noState := v.mainTreeViewer(committed).NoState()
-	voteCountLE, err := noState.Get(voteCountKey)
-	if errors.Is(err, db.ErrKeyNotFound) {
-		return 0, nil
-	} else if err != nil {
-		return 0, err
-	}
-	return binary.LittleEndian.Uint64(voteCountLE), nil
-}
-
-// voteCountInc increases by 1 the global vote count.
-func (v *State) voteCountInc() error {
-	noState := v.Tx.NoState()
-	voteCountLE, err := noState.Get(voteCountKey)
-	if errors.Is(err, db.ErrKeyNotFound) {
-		voteCountLE = make([]byte, 8)
-	} else if err != nil {
-		return err
-	}
-	voteCount := binary.LittleEndian.Uint64(voteCountLE)
-	voteCount++
-	binary.LittleEndian.PutUint64(voteCountLE, voteCount)
-	return noState.Set(voteCountKey, voteCountLE)
-}
-
-// AddVote adds a new vote to a process and call the even listeners to OnVote.
-// This method does not check if the vote already exist!
-func (v *State) AddVote(vote *models.Vote, voterID VoterID) error {
-	vid, err := v.voteID(vote.ProcessId, vote.Nullifier)
-	if err != nil {
-		return err
-	}
-	// save block number
-	vote.Height = v.CurrentHeight()
-	voteBytes, err := proto.Marshal(vote)
-	if err != nil {
-		return fmt.Errorf("cannot marshal vote: %w", err)
-	}
-	// TO-DO (pau): Why are we storing processID and nullifier?
-	sdbVote := models.StateDBVote{
-		VoteHash:  ethereum.HashRaw(voteBytes),
-		ProcessId: vote.ProcessId,
-		Nullifier: vote.Nullifier,
-	}
-	sdbVoteBytes, err := proto.Marshal(&sdbVote)
-	if err != nil {
-		return fmt.Errorf("cannot marshal sdbVote: %w", err)
-	}
-	v.Tx.Lock()
-	err = func() error {
-		treeCfg := StateChildTreeCfg(ChildTreeVotes)
-		if err := v.Tx.DeepAdd(vid, sdbVoteBytes,
-			StateTreeCfg(TreeProcess), treeCfg.WithKey(vote.ProcessId)); err != nil {
-			return err
-		}
-		return v.voteCountInc()
-	}()
-	v.Tx.Unlock()
-	if err != nil {
-		return err
-	}
-	for _, l := range v.eventListeners {
-		l.OnVote(vote, voterID, v.TxCounter())
-	}
-	return nil
-}
-
-// NOTE(Edu): Changed this from byte(processID+nullifier) to
-// hash(processID+nullifier) to allow using it as a key in Arbo tree.
-// voteID = hash(processID+nullifier)
-func (v *State) voteID(pid, nullifier []byte) ([]byte, error) {
-	if len(pid) != types.ProcessIDsize {
-		return nil, fmt.Errorf("wrong processID size %d", len(pid))
-	}
-	if len(nullifier) != types.VoteNullifierSize {
-		return nil, fmt.Errorf("wrong nullifier size %d", len(nullifier))
-	}
-	vid := sha256.New()
-	vid.Write(pid)
-	vid.Write(nullifier)
-	return vid.Sum(nil), nil
-}
-
-// Envelope returns the hash of a stored vote if exists.
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) Envelope(processID, nullifier []byte, committed bool) (_ []byte, err error) {
-	vid, err := v.voteID(processID, nullifier)
-	if err != nil {
-		return nil, err
-	}
-	if !committed {
-		// acquire a write lock, since DeepSubTree will create some temporary trees in memory
-		// that might be read concurrently by DeliverTx path during block commit, leading to race #581
-		// https://github.com/vocdoni/vocdoni-node/issues/581
-		v.Tx.Lock()
-		defer v.Tx.Unlock()
-	}
-	treeCfg := StateChildTreeCfg(ChildTreeVotes)
-	votesTree, err := v.mainTreeViewer(committed).DeepSubTree(
-		StateTreeCfg(TreeProcess), treeCfg.WithKey(processID))
-	if errors.Is(err, arbo.ErrKeyNotFound) {
-		return nil, ErrProcessNotFound
-	} else if err != nil {
-		return nil, err
-	}
-	sdbVoteBytes, err := votesTree.Get(vid)
-	if errors.Is(err, arbo.ErrKeyNotFound) {
-		return nil, ErrVoteDoesNotExist
-	} else if err != nil {
-		return nil, err
-	}
-	var sdbVote models.StateDBVote
-	if err := proto.Unmarshal(sdbVoteBytes, &sdbVote); err != nil {
-		return nil, fmt.Errorf("cannot unmarshal sdbVote: %w", err)
-	}
-	return sdbVote.VoteHash, nil
-}
-
-// EnvelopeExists returns true if the envelope identified with voteID exists
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) EnvelopeExists(processID, nullifier []byte, committed bool) (bool, error) {
-	_, err := v.Envelope(processID, nullifier, committed)
-	if errors.Is(err, ErrProcessNotFound) {
-		return false, nil
-	} else if errors.Is(err, ErrVoteDoesNotExist) {
-		return false, nil
-	} else if err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-// iterateVotes iterates fn over state tree entries with the processID prefix.
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) iterateVotes(processID []byte,
-	fn func(vid []byte, sdbVote *models.StateDBVote) bool, committed bool) error {
-	if !committed {
-		v.Tx.RLock()
-		defer v.Tx.RUnlock()
-	}
-	treeCfg := StateChildTreeCfg(ChildTreeVotes)
-	votesTree, err := v.mainTreeViewer(committed).DeepSubTree(
-		StateTreeCfg(TreeProcess), treeCfg.WithKey(processID))
-	if err != nil {
-		return err
-	}
-	var callbackErr error
-	if err := votesTree.Iterate(func(key, value []byte) bool {
-		var sdbVote models.StateDBVote
-		if err := proto.Unmarshal(value, &sdbVote); err != nil {
-			callbackErr = err
-			return true
-		}
-		return fn(key, &sdbVote)
-	}); err != nil {
-		return err
-	}
-	if callbackErr != nil {
-		return callbackErr
-	}
-	return nil
-}
-
-// CountVotes returns the number of votes registered for a given process id
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) CountVotes(processID []byte, committed bool) uint32 {
-	var count uint32
-	// TODO: Once statedb.TreeView.Size() works, replace this by that.
-	v.iterateVotes(processID, func(vid []byte, sdbVote *models.StateDBVote) bool {
-		count++
-		return false
-	}, committed)
-	return count
-}
-
-// EnvelopeList returns a list of registered envelopes nullifiers given a processId
-// When committed is false, the operation is executed also on not yet commited
-// data from the currently open StateDB transaction.
-// When committed is true, the operation is executed on the last commited version.
-func (v *State) EnvelopeList(processID []byte, from, listSize int,
-	committed bool) (nullifiers [][]byte) {
-	idx := 0
-	v.iterateVotes(processID, func(vid []byte, sdbVote *models.StateDBVote) bool {
-		if idx >= from+listSize {
-			return true
-		}
-		if idx >= from {
-			nullifiers = append(nullifiers, sdbVote.Nullifier)
-		}
-		idx++
-		return false
-	}, committed)
-	return nullifiers
 }
 
 // pathProcessIDsByStartBlock is the db path used to store ProcessIDs indexed
@@ -1144,23 +562,4 @@ func (v *State) TxCounterAdd() {
 // TxCounter returns the current tx count
 func (v *State) TxCounter() int32 {
 	return atomic.LoadInt32(&v.txCounter)
-}
-
-// VerifyTreasurer checks is an address is the treasurer and the
-// nonce provided is the expected one
-func (v *State) VerifyTreasurer(addr common.Address, txNonce uint32) error {
-	// get treasurer
-	treasurer, err := v.Treasurer(false)
-	if err != nil {
-		return fmt.Errorf("cannot check authorization")
-	}
-	log.Debugf("got treasurer addr %x", treasurer.Address)
-	if !bytes.Equal(addr.Bytes(), treasurer.Address) {
-		return fmt.Errorf("not authorized for executing admin transactions")
-	}
-	// check treasurer account
-	if treasurer.Nonce != txNonce {
-		return ErrAccountNonceInvalid
-	}
-	return nil
 }
