+++ vochain/state_account.go
@@ -87,69 +87,6 @@ func (a *Account) DelDelegate(addr common.Address) error {
 	return nil
 }
 
-// TransferBalance transfers balance from origin address to destination address,
-// and updates the state with the new values (including nonce).
-// If origin address acc is not enough, ErrNotEnoughBalance is returned.
-func (v *State) TransferBalance(from, to common.Address, amount uint64) error {
-	accFrom, err := v.GetAccount(from, false)
-	if err != nil {
-		return err
-	}
-	if accFrom == nil {
-		return ErrAccountNotExist
-	}
-	accTo, err := v.GetAccount(to, false)
-	if err != nil {
-		return err
-	}
-	if accTo == nil {
-		return ErrAccountNotExist
-	}
-	if err := accFrom.Transfer(accTo, amount); err != nil {
-		return err
-	}
-	log.Debugf("transferring %d tokens from %s to %s", amount, from.String(), to.String())
-	if err := v.SetAccount(from, accFrom); err != nil {
-		return err
-	}
-	if err := v.SetAccount(to, accTo); err != nil {
-		return err
-	}
-	for _, l := range v.eventListeners {
-		l.OnTransferTokens(from.Bytes(), to.Bytes(), amount)
-	}
-	return nil
-}
-
-// MintBalance increments the existing acc of address by amount
-func (v *State) MintBalance(address common.Address, amount uint64) error {
-	if amount == 0 {
-		return fmt.Errorf("cannot mint a zero amount balance")
-	}
-	acc, err := v.GetAccount(address, false)
-	if err != nil {
-		return fmt.Errorf("mintBalance: %w", err)
-	}
-	if acc == nil {
-		return ErrAccountNotExist
-	}
-	if acc.Balance+amount <= acc.Balance {
-		return ErrBalanceOverflow
-	}
-	acc.Balance += amount
-	log.Debugf("minting %d tokens to account %s", amount, address.String())
-	return v.SetAccount(address, acc)
-}
-
-// BurnTxCost burns the cost of a transaction
-// if cost is set to 0 just return
-func (v *State) BurnTxCost(from common.Address, cost uint64) error {
-	if cost != 0 {
-		return v.TransferBalance(from, BurnAddress, cost)
-	}
-	return nil
-}
-
 // GetAccount retrieves the Account for an address.
 // Returns a nil account and no error if the account does not exist.
 // Committed is relative to the state on which the function is executed.
@@ -235,29 +172,6 @@ func (v *State) IncrementAccountProcessIndex(accountAddress common.Address) erro
 	acc.ProcessIndex++
 	log.Debugf("setting account %s process index to %d", accountAddress.String(), acc.ProcessIndex)
 	return v.SetAccount(accountAddress, acc)
-}
-
-// ConsumeFaucetPayload consumes a given faucet payload storing
-// its key to the FaucetNonce tree so it can only be used once
-func (v *State) ConsumeFaucetPayload(from common.Address, faucetPayload *models.FaucetPayload) error {
-	// check faucet payload
-	if faucetPayload == nil {
-		return fmt.Errorf("invalid faucet payload")
-	}
-	// store faucet identifier
-	b := make([]byte, 8)
-	binary.LittleEndian.PutUint64(b, faucetPayload.Identifier)
-	keyHash := ethereum.HashRaw(append(from.Bytes(), b...))
-	if err := v.SetFaucetNonce(keyHash); err != nil {
-		return err
-	}
-	log.Debugf("consuming faucet payload created by %s with amount %d and identifier %d (keyHash: %x)",
-		from.String(),
-		faucetPayload.Amount,
-		faucetPayload.Identifier,
-		keyHash,
-	)
-	return nil
 }
 
 // CreateAccountTxCheck checks if an account creation tx is valid
@@ -456,14 +370,7 @@ func (v *State) SetAccount(accountAddress common.Address, account *Account) erro
 	if err != nil {
 		return err
 	}
-	log.Debugf(`setAccount: {
-			address %s,
-			nonce %d,
-			infoURI %s,
-			balance: %d,
-			delegates: %+v,
-			processIndex: %d
-		}`,
+	log.Debugf("setAccount: address %s, nonce %d, infoURI %s, balance: %d, delegates: %+v, processIndex: %d",
 		accountAddress.String(),
 		account.Nonce,
 		account.InfoURI,
@@ -526,175 +433,6 @@ func (v *State) BurnTxCostIncrementNonce(accountAddress common.Address, txType m
 	return nil
 }
 
-// MintTokensTxCheck checks if a given MintTokensTx and its data are valid
-func MintTokensTxCheck(vtx *models.Tx, txBytes, signature []byte, state *State) error {
-	if vtx == nil || txBytes == nil || signature == nil || state == nil {
-		return fmt.Errorf("invalid parameters")
-	}
-	tx := vtx.GetMintTokens()
-	if tx == nil {
-		return fmt.Errorf("invalid tx")
-	}
-	if tx.Value <= 0 {
-		return fmt.Errorf("invalid value")
-	}
-	if len(tx.To) == 0 {
-		return fmt.Errorf("invalid To address")
-	}
-	pubKey, err := ethereum.PubKeyFromSignature(txBytes, signature)
-	if err != nil {
-		return fmt.Errorf("cannot extract public key from signature: %w", err)
-	}
-	txSenderAddress, err := ethereum.AddrFromPublicKey(pubKey)
-	if err != nil {
-		return fmt.Errorf("cannot extract address from public key: %w", err)
-	}
-	treasurer, err := state.Treasurer(false)
-	if err != nil {
-		return err
-	}
-	treasurerAddress := common.BytesToAddress(treasurer.Address)
-	if treasurerAddress != txSenderAddress {
-		return fmt.Errorf(
-			"address recovered not treasurer: expected %s got %s",
-			treasurerAddress.String(),
-			txSenderAddress.String(),
-		)
-	}
-	if tx.Nonce != treasurer.Nonce {
-		return fmt.Errorf("invalid nonce %d, expected: %d", tx.Nonce, treasurer.Nonce)
-	}
-	toAddr := common.BytesToAddress(tx.To)
-	toAcc, err := state.GetAccount(toAddr, false)
-	if err != nil {
-		return fmt.Errorf("cannot get to account: %w", err)
-	}
-	if toAcc == nil {
-		return ErrAccountNotExist
-	}
-	return nil
-}
-
-// SendTokensTxCheck checks if a given SendTokensTx and its data are valid
-func SendTokensTxCheck(vtx *models.Tx, txBytes, signature []byte, state *State) error {
-	if vtx == nil || signature == nil || txBytes == nil || state == nil {
-		return fmt.Errorf("invalid parameters")
-	}
-	tx := vtx.GetSendTokens()
-	if tx == nil {
-		return fmt.Errorf("invalid tx")
-	}
-	if tx.Value == 0 {
-		return fmt.Errorf("invalid value")
-	}
-	if len(tx.From) == 0 {
-		return fmt.Errorf("invalid from address")
-	}
-	if len(tx.To) == 0 {
-		return fmt.Errorf("invalid to address")
-	}
-	pubKey, err := ethereum.PubKeyFromSignature(txBytes, signature)
-	if err != nil {
-		return fmt.Errorf("cannot extract public key from signature: %w", err)
-	}
-	txSenderAddress, err := ethereum.AddrFromPublicKey(pubKey)
-	if err != nil {
-		return fmt.Errorf("cannot extract address from public key: %w", err)
-	}
-	txFromAddress := common.BytesToAddress(tx.From)
-	if txFromAddress != txSenderAddress {
-		return fmt.Errorf("from (%s) field and extracted signature (%s) mismatch",
-			txFromAddress.String(),
-			txSenderAddress.String(),
-		)
-	}
-	txToAddress := common.BytesToAddress(tx.To)
-	toTxAccount, err := state.GetAccount(txToAddress, false)
-	if err != nil {
-		return fmt.Errorf("cannot get to account: %w", err)
-	}
-	if toTxAccount == nil {
-		return ErrAccountNotExist
-	}
-	acc, err := state.GetAccount(txSenderAddress, false)
-	if err != nil {
-		return fmt.Errorf("cannot get from account: %w", err)
-	}
-	if acc == nil {
-		return ErrAccountNotExist
-	}
-	if tx.Nonce != acc.Nonce {
-		return fmt.Errorf("invalid nonce, expected %d got %d", acc.Nonce, tx.Nonce)
-	}
-	cost, err := state.TxCost(models.TxType_SEND_TOKENS, false)
-	if err != nil {
-		return err
-	}
-	if (tx.Value + cost) > acc.Balance {
-		return ErrNotEnoughBalance
-	}
-	return nil
-}
-
-// SetAccountDelegateTxCheck checks if a SetAccountDelegateTx and its data are valid
-func SetAccountDelegateTxCheck(vtx *models.Tx, txBytes, signature []byte, state *State) error {
-	if vtx == nil || signature == nil || txBytes == nil || state == nil {
-		return fmt.Errorf("invalid parameters")
-	}
-	tx := vtx.GetSetAccount()
-	if tx == nil {
-		return fmt.Errorf("invalid tx")
-	}
-	if tx.Txtype != models.TxType_ADD_DELEGATE_FOR_ACCOUNT &&
-		tx.Txtype != models.TxType_DEL_DELEGATE_FOR_ACCOUNT {
-		return fmt.Errorf("invalid tx type")
-	}
-	if tx.Nonce == nil {
-		return fmt.Errorf("invalid nonce")
-	}
-	if len(tx.Delegates) == 0 {
-		return fmt.Errorf("invalid delegates")
-	}
-	txSenderAddress, txSenderAccount, err := state.AccountFromSignature(txBytes, signature)
-	if err != nil {
-		return err
-	}
-	if err := checkDuplicateDelegates(tx.Delegates, txSenderAddress); err != nil {
-		return fmt.Errorf("checkDuplicateDelegates: %w", err)
-	}
-	if tx.GetNonce() != txSenderAccount.Nonce {
-		return fmt.Errorf("invalid nonce, expected %d got %d", txSenderAccount.Nonce, tx.Nonce)
-	}
-	cost, err := state.TxCost(tx.Txtype, false)
-	if err != nil {
-		return fmt.Errorf("cannot get tx cost: %w", err)
-	}
-	if txSenderAccount.Balance < cost {
-		return ErrNotEnoughBalance
-	}
-	switch tx.Txtype {
-	case models.TxType_ADD_DELEGATE_FOR_ACCOUNT:
-		for _, delegate := range tx.Delegates {
-			delegateAddress := common.BytesToAddress(delegate)
-			if txSenderAccount.IsDelegate(delegateAddress) {
-				return fmt.Errorf("delegate %s already exists", delegateAddress.String())
-			}
-		}
-		return nil
-	case models.TxType_DEL_DELEGATE_FOR_ACCOUNT:
-		for _, delegate := range tx.Delegates {
-			delegateAddress := common.BytesToAddress(delegate)
-			if !txSenderAccount.IsDelegate(delegateAddress) {
-				return fmt.Errorf("delegate %s does not exist", delegateAddress.String())
-			}
-		}
-		return nil
-	default:
-		// should never happen
-		return fmt.Errorf("invalid tx type")
-	}
-}
-
 // SetAccountDelegate sets a set of delegates for a given account
 func (v *State) SetAccountDelegate(accountAddr common.Address,
 	delegateAddrs [][]byte,
@@ -726,91 +464,6 @@ func (v *State) SetAccountDelegate(accountAddr common.Address,
 	default:
 		return fmt.Errorf("invalid tx type")
 	}
-}
-
-// CollectFaucetTxCheck checks if a CollectFaucetTx and its data are valid
-func CollectFaucetTxCheck(vtx *models.Tx, txBytes, signature []byte, state *State) error {
-	if vtx == nil || signature == nil || txBytes == nil || state == nil {
-		return fmt.Errorf("invalid parameters")
-	}
-	tx := vtx.GetCollectFaucet()
-	if tx == nil {
-		return fmt.Errorf("invalid tx")
-	}
-	faucetPkg := tx.GetFaucetPackage()
-	if faucetPkg == nil {
-		return fmt.Errorf("nil faucet package")
-	}
-	if faucetPkg.Signature == nil {
-		return fmt.Errorf("invalid faucet package signature")
-	}
-	if faucetPkg.Payload == nil {
-		return fmt.Errorf("invalid faucet package payload")
-	}
-	faucetPayload := &models.FaucetPayload{}
-	if err := proto.Unmarshal(tx.FaucetPackage.Payload, faucetPayload); err != nil {
-		return fmt.Errorf("could not unmarshal faucet package: %w", err)
-	}
-	if faucetPayload.Amount == 0 {
-		return fmt.Errorf("invalid faucet package payload amount")
-	}
-	if len(faucetPayload.To) == 0 {
-		return fmt.Errorf("invalid faucet package payload to")
-	}
-	payloadToAddress := common.BytesToAddress(faucetPayload.To)
-	pubKey, err := ethereum.PubKeyFromSignature(txBytes, signature)
-	if err != nil {
-		return fmt.Errorf("cannot extract public key from signature: %w", err)
-	}
-	txSenderAddress, err := ethereum.AddrFromPublicKey(pubKey)
-	if err != nil {
-		return fmt.Errorf("cannot extract address from public key: %w", err)
-	}
-	if txSenderAddress != payloadToAddress {
-		return fmt.Errorf("txSender %s and faucet payload to %s mismatch",
-			txSenderAddress,
-			payloadToAddress,
-		)
-	}
-	txSenderAccount, err := state.GetAccount(txSenderAddress, false)
-	if err != nil {
-		return fmt.Errorf("cannot check if account %s exists: %w", txSenderAddress.String(), err)
-	}
-	if txSenderAccount == nil {
-		return ErrAccountNotExist
-	}
-	if txSenderAccount.Nonce != tx.Nonce {
-		return fmt.Errorf("invalid nonce")
-	}
-	fromAddr, err := ethereum.AddrFromSignature(tx.FaucetPackage.Payload, tx.FaucetPackage.Signature)
-	if err != nil {
-		return fmt.Errorf("cannot extract address from faucet package signature: %w", err)
-	}
-	b := make([]byte, 8)
-	binary.LittleEndian.PutUint64(b, faucetPayload.Identifier)
-	keyHash := ethereum.HashRaw(append(fromAddr.Bytes(), b...))
-	used, err := state.FaucetNonce(keyHash, false)
-	if err != nil {
-		return fmt.Errorf("cannot check faucet nonce: %w", err)
-	}
-	if used {
-		return fmt.Errorf("faucet payload already used")
-	}
-	issuerAcc, err := state.GetAccount(fromAddr, false)
-	if err != nil {
-		return fmt.Errorf("cannot get faucet account: %w", err)
-	}
-	if issuerAcc == nil {
-		return fmt.Errorf("the account signing the faucet payload does not exist")
-	}
-	cost, err := state.TxCost(models.TxType_COLLECT_FAUCET, false)
-	if err != nil {
-		return fmt.Errorf("cannot get %s tx cost: %w", models.TxType_COLLECT_FAUCET, err)
-	}
-	if issuerAcc.Balance < faucetPayload.Amount+cost {
-		return fmt.Errorf("faucet does not have enough balance %d, required %d", issuerAcc.Balance, faucetPayload.Amount+cost)
-	}
-	return nil
 }
 
 // GenerateFaucetPackage generates a faucet package
