+++ vochain/state_process.go
@@ -5,19 +5,18 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/vocdoni/arbo"
-	"go.vocdoni.io/dvote/crypto/zk/artifacts"
-	"go.vocdoni.io/dvote/log"
 	"go.vocdoni.io/dvote/statedb"
 	"go.vocdoni.io/dvote/types"
 	"go.vocdoni.io/proto/build/go/models"
 	"google.golang.org/protobuf/proto"
 )
 
-var emptyVotesRoot = make([]byte, StateChildTreeCfg(ChildTreeVotes).HashFunc().Len())
-var emptyCensusRoot = make([]byte, StateChildTreeCfg(ChildTreeCensus).HashFunc().Len())
-var emptyPreRegisterNullifiersRoot = make([]byte, StateChildTreeCfg(ChildTreePreRegisterNullifiers).HashFunc().Len())
+var (
+	emptyVotesRoot                 = make([]byte, StateChildTreeCfg(ChildTreeVotes).HashFunc().Len())
+	emptyCensusRoot                = make([]byte, StateChildTreeCfg(ChildTreeCensus).HashFunc().Len())
+	emptyPreRegisterNullifiersRoot = make([]byte, StateChildTreeCfg(ChildTreePreRegisterNullifiers).HashFunc().Len())
+)
 
 // AddProcess adds a new process to the vochain.  Adding a process with a
 // ProcessId that already exists will return an error.
@@ -338,9 +337,7 @@ func (v *State) SetProcessResults(pid []byte, result *models.ProcessResult, comm
 		}
 		// Call event listeners
 		for _, l := range v.eventListeners {
-			if err := l.OnProcessResults(process.ProcessId, result, v.TxCounter()); err != nil {
-				log.Warnf("onProcessResults callback error: %v", err)
-			}
+			l.OnProcessResults(process.ProcessId, result, v.TxCounter())
 		}
 	}
 	return nil
@@ -402,227 +399,4 @@ func (v *State) SetProcessCensus(pid, censusRoot []byte, censusURI string, commi
 	}
 
 	return nil
-}
-
-// NewProcessTxCheck is an abstraction of ABCI checkTx for creating a new process
-func (app *BaseApplication) NewProcessTxCheck(vtx *models.Tx, txBytes,
-	signature []byte, state *State) (*models.Process, common.Address, error) {
-	tx := vtx.GetNewProcess()
-	if tx.Process == nil {
-		return nil, common.Address{}, fmt.Errorf("process data is empty")
-	}
-	// basic required fields check
-	if tx.Process.VoteOptions == nil || tx.Process.EnvelopeType == nil || tx.Process.Mode == nil {
-		return nil, common.Address{}, fmt.Errorf("missing required fields (voteOptions, envelopeType or processMode)")
-	}
-	if tx.Process.VoteOptions.MaxCount == 0 {
-		return nil, common.Address{}, fmt.Errorf("missing vote maxCount parameter")
-	}
-	// check signature available
-	if signature == nil || tx == nil || txBytes == nil {
-		return nil, common.Address{}, fmt.Errorf("missing signature or new process transaction")
-	}
-	// start and block count sanity check
-	// if startBlock is zero or one, the process will be enabled on the next block
-	if tx.Process.StartBlock == 0 || tx.Process.StartBlock == 1 {
-		tx.Process.StartBlock = state.CurrentHeight() + 1
-	} else if tx.Process.StartBlock < state.CurrentHeight() {
-		return nil, common.Address{}, fmt.Errorf(
-			"cannot add process with start block lower than or equal to the current height")
-	}
-	if tx.Process.BlockCount <= 0 {
-		return nil, common.Address{}, fmt.Errorf(
-			"cannot add process with duration lower than or equal to the current height")
-	}
-	// check tx cost
-	cost, err := state.TxCost(models.TxType_NEW_PROCESS, false)
-	if err != nil {
-		return nil, common.Address{}, fmt.Errorf("cannot get NewProcessTx transaction cost: %w", err)
-	}
-	addr, acc, err := state.AccountFromSignature(txBytes, signature)
-	if err != nil {
-		return nil, common.Address{}, err
-	}
-	if addr == nil {
-		return nil, common.Address{}, fmt.Errorf("cannot get account from signature, nil result")
-	}
-	// check balance and nonce
-	if acc.Balance < cost {
-		return nil, common.Address{}, ErrNotEnoughBalance
-	}
-	if acc.Nonce != tx.Nonce {
-		return nil, common.Address{}, ErrAccountNonceInvalid
-	}
-
-	isOracle, err := state.IsOracle(*addr)
-	if err != nil {
-		return nil, common.Address{}, err
-	}
-
-	// check if process entityID matches tx sender
-	if !bytes.Equal(tx.Process.EntityId, addr.Bytes()) && !isOracle {
-		// if not oracle check delegate
-		entityAddress := common.BytesToAddress(tx.Process.EntityId)
-		entityAccount, err := state.GetAccount(entityAddress, false)
-		if err != nil {
-			return nil, common.Address{}, fmt.Errorf(
-				"cannot get entity account for checking if the sender is a delegate: %w", err,
-			)
-		}
-		if entityAccount == nil {
-			return nil, common.Address{}, ErrAccountNotExist
-		}
-		if !entityAccount.IsDelegate(*addr) {
-			return nil, common.Address{}, fmt.Errorf(
-				"unauthorized to create a new process, recovered addr is %s", addr.Hex())
-		}
-	}
-
-	// if no Oracle, build the processID (Oracles are allowed to use any processID)
-	if !isOracle || tx.Process.ProcessId == nil {
-		// if Oracle but processID empty, switch the entityID temporary to the Oracle address
-		// this way we ensure the account creating the process exists
-		entityID := tx.Process.EntityId
-		if isOracle {
-			tx.Process.EntityId = addr.Bytes()
-		}
-		pid, err := app.BuildProcessID(tx.Process)
-		if err != nil {
-			return nil, common.Address{}, fmt.Errorf("cannot build processID: %w", err)
-		}
-		tx.Process.ProcessId = pid.Marshal()
-		// restore original entityID
-		tx.Process.EntityId = entityID
-	}
-
-	// check if process already exists
-	_, err = state.Process(tx.Process.ProcessId, false)
-	if err == nil {
-		return nil, common.Address{}, fmt.Errorf("process with id (%x) already exists", tx.Process.ProcessId)
-	}
-
-	// check valid/implemented process types
-	// pre-regiser and anonymous must be either both enabled or disabled, as
-	// we only support a single scenario of pre-register + anonymous.
-	if tx.Process.Mode.PreRegister != tx.Process.EnvelopeType.Anonymous {
-		return nil, common.Address{}, fmt.Errorf("pre-register mode only supported " +
-			"with anonymous envelope type and viceversa")
-	}
-	if tx.Process.Mode.PreRegister &&
-		(tx.Process.MaxCensusSize == nil || tx.Process.GetMaxCensusSize() <= 0) {
-		return nil, common.Address{}, fmt.Errorf("pre-register mode requires setting " +
-			"maxCensusSize to be > 0")
-	}
-	if tx.Process.Mode.PreRegister && tx.Process.EnvelopeType.Anonymous {
-		var circuits []artifacts.CircuitConfig
-		if genesis, ok := Genesis[app.chainID]; ok {
-			circuits = genesis.CircuitsConfig
-		} else {
-			log.Warn("Using dev network genesis CircuitsConfig")
-			circuits = Genesis["dev"].CircuitsConfig
-		}
-		if len(circuits) == 0 {
-			return nil, common.Address{}, fmt.Errorf("no circuit configs in the %v genesis", app.chainID)
-		}
-		if tx.Process.MaxCensusSize == nil {
-			return nil, common.Address{}, fmt.Errorf("maxCensusSize is not provided")
-		}
-		if tx.Process.GetMaxCensusSize() > uint64(circuits[len(circuits)-1].Parameters[0]) {
-			return nil, common.Address{}, fmt.Errorf("maxCensusSize for anonymous envelope "+
-				"cannot be bigger than the parameter for the biggest circuit (%v)",
-				circuits[len(circuits)-1].Parameters[0])
-		}
-	}
-
-	// TODO: Enable support for PreRegiser without Anonymous.  Figure out
-	// all the required changes to support a process with a rolling census
-	// that is not Anonymous.
-	if tx.Process.EnvelopeType.Serial {
-		return nil, common.Address{}, fmt.Errorf("serial process not yet implemented")
-	}
-
-	if tx.Process.EnvelopeType.EncryptedVotes || tx.Process.EnvelopeType.Anonymous {
-		// We consider the zero value as nil for security
-		tx.Process.EncryptionPublicKeys = make([]string, types.KeyKeeperMaxKeyIndex)
-		tx.Process.EncryptionPrivateKeys = make([]string, types.KeyKeeperMaxKeyIndex)
-	}
-	return tx.Process, *addr, nil
-}
-
-// SetProcessTxCheck is an abstraction of ABCI checkTx for canceling an existing process
-func SetProcessTxCheck(vtx *models.Tx, txBytes, signature []byte, state *State) (common.Address, error) {
-	tx := vtx.GetSetProcess()
-	// check signature available
-	if signature == nil || tx == nil || txBytes == nil {
-		return common.Address{}, fmt.Errorf("missing signature on setProcess transaction")
-	}
-	// get tx cost
-	cost, err := state.TxCost(tx.Txtype, false)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("cannot get %s transaction cost: %w", tx.Txtype.String(), err)
-	}
-	addr, acc, err := state.AccountFromSignature(txBytes, signature)
-	if err != nil {
-		return common.Address{}, err
-	}
-	// check balance and nonce
-	if acc.Balance < cost {
-		return common.Address{}, ErrNotEnoughBalance
-	}
-	if acc.Nonce != tx.Nonce {
-		return common.Address{}, ErrAccountNonceInvalid
-	}
-	// get process
-	process, err := state.Process(tx.ProcessId, false)
-	if err != nil {
-		return common.Address{}, fmt.Errorf("cannot get process %x: %w", tx.ProcessId, err)
-	}
-	// check process entityID matches tx sender
-	isOracle := false
-	if !bytes.Equal(process.EntityId, addr.Bytes()) {
-		// Check if the transaction comes from an oracle
-		// Oracles can create processes with any entityID
-		isOracle, err = state.IsOracle(*addr)
-		if err != nil {
-			return common.Address{}, err
-		}
-		if !isOracle {
-			// check if delegate
-			entityIDAddress := common.BytesToAddress(process.EntityId)
-			entityIDAccount, err := state.GetAccount(entityIDAddress, true)
-			if err != nil {
-				return common.Address{}, fmt.Errorf(
-					"cannot get entityID account for checking if the sender is a delegate: %w", err,
-				)
-			}
-			if !entityIDAccount.IsDelegate(*addr) {
-				return common.Address{}, fmt.Errorf(
-					"unauthorized to set process status, recovered addr is %s", addr.Hex(),
-				)
-			} // is delegate
-		} // is oracle
-	}
-	switch tx.Txtype {
-	case models.TxType_SET_PROCESS_RESULTS:
-		if !isOracle {
-			return common.Address{}, fmt.Errorf("only oracles can execute set process results transaction")
-		}
-		if acc.Balance < cost {
-			return common.Address{}, ErrNotEnoughBalance
-		}
-		if acc.Nonce != tx.Nonce {
-			return common.Address{}, ErrAccountNonceInvalid
-		}
-		results := tx.GetResults()
-		if !bytes.Equal(results.OracleAddress, addr.Bytes()) {
-			return common.Address{}, fmt.Errorf("cannot set results, oracle address provided in results does not match")
-		}
-		return *addr, state.SetProcessResults(process.ProcessId, results, false)
-	case models.TxType_SET_PROCESS_STATUS:
-		return *addr, state.SetProcessStatus(process.ProcessId, tx.GetStatus(), false)
-	case models.TxType_SET_PROCESS_CENSUS:
-		return *addr, state.SetProcessCensus(process.ProcessId, tx.GetCensusRoot(), tx.GetCensusURI(), false)
-	default:
-		return common.Address{}, fmt.Errorf("unknown setProcess tx type: %s", tx.Txtype)
-	}
 }
