+++ vochain/state_test.go
@@ -158,14 +158,11 @@ func (l *Listener) OnProcessStatusChange(pid []byte, status models.ProcessStatus
 func (l *Listener) OnCancel(pid []byte, txIndex int32)                                           {}
 func (l *Listener) OnProcessKeys(pid []byte, encryptionPub string, txIndex int32)                {}
 func (l *Listener) OnRevealKeys(pid []byte, encryptionPriv string, txIndex int32)                {}
-func (l *Listener) OnProcessResults(pid []byte, results *models.ProcessResult, txIndex int32) error {
-	return nil
+func (l *Listener) OnProcessResults(pid []byte, results *models.ProcessResult, txIndex int32) {
 }
-func (l *Listener) OnSetAccount(addr []byte, account *Account) error {
-	return nil
+func (l *Listener) OnSetAccount(addr []byte, account *Account) {
 }
-func (l *Listener) OnTransferTokens(from, to []byte, amount uint64) error {
-	return nil
+func (l *Listener) OnTransferTokens(from, to []byte, amount uint64) {
 }
 func (l *Listener) OnProcessesStart(pids [][]byte) {
 	l.processStart = append(l.processStart, pids)
